move status helper functions
to implement them in the order details screen as well, enhance the
 status display element
