+++ components/OrderCard.tsx
@@ -5,11 +5,13 @@ import { Ionicons } from '@expo/vector-icons';
 import { useOrder } from '@/context/orderContext'
 import firebaseService from '@/handlers/firebaseService'
 import { useUser } from '@/context/userContext';
+import { useOrderStatus } from '@/hooks/useOrderStatus';
 
 const OrderCard = ({ order, onPress }) => {
     const { date, total, shops, items } = order;
     const { selectedOrder, setSelectedOrder } = useOrder()
     const { userData } = useUser();
+    const { getStatusColor, getStatusText } = useOrderStatus();
 
     const [customerName, setCustomerName] = useState<string>('unknown');
 
@@ -24,40 +26,6 @@ const OrderCard = ({ order, onPress }) => {
         });
         return () => { isMounted = false; };
     }, [order.customerId]);
-
-    const getStatusColor = (status: string) => {
-        switch (status) {
-            case 'pending':
-                return '#FF9800';
-            case 'preparing':
-                return '#9C27B0';
-            case 'ready':
-                return '#00bfff';
-            case 'in-delivery':
-                return 'orchid';
-            case 'delivered':
-                return '#4CAF50';
-            default:
-                return '#4f6549ff';
-        }
-    };
-
-    const getStatusText = (status: string) => {
-        switch (status) {
-            case 'pending':
-                return 'Order Received';
-            case 'preparing':
-                return 'Being Prepared';
-            case 'ready':
-                return 'Ready for Pickup';
-            case 'in-delivery':
-                return 'Out for Delivery';
-            case 'completed':
-                return 'Delivered';
-            default:
-                return status;
-        }
-    };
 
     const handleStatusPress = (order) => {
         alert(`ORDER TOTAL: $${order.total}`)
